# perf(ext/headers): use .push loop instead of spread operator