@bartlomieju 2023-08-09
@marcosc90 very nice, how do the results look like if you increase number of headers to 8-10?