@marcosc90 2023-08-09
@bartlomieju it's always faster, here's the bench for 8 headers.

```
const headers = new Headers({
  "Content-Type": "application/json",
  "Date": "Wed, 09 Aug 2023 11:37:57 GMT",
  "X-Deno": "Deno",
  "Powered-By": "Deno",
  "Content-Encoding": "gzip",
  "Set-Cookie": "__Secure-ID=123; Secure; Domain=example.com",
  "Content-Length": "150",
  "Vary": "Accept-Encoding, Accept, X-Requested-With"
});
```
**This patch**

```
benchmark             time (avg)        iter/s             (min … max)       p75       p99      p995
---------------------------------------------------------------------- -----------------------------
Headers iterator      10.59 µs/iter      94,446.5   (9.64 µs … 107.35 µs)  10.39 µs  14.51 µs  49.75 µs
```

**Main**
```
benchmark             time (avg)        iter/s             (min … max)       p75       p99      p995
---------------------------------------------------------------------- -----------------------------
Headers iterator      12.03 µs/iter      83,097.9  (10.97 µs … 243.54 µs)  11.76 µs  16.97 µs  53.33 µs
```