@marcosc90 2023-08-09
@bartlomieju got much better results with the latest push.

2 headers
```
benchmark             time (avg)        iter/s             (min … max)       p75       p99      p995
---------------------------------------------------------------------- -----------------------------
Headers iterator        1.3 µs/iter     771,516.2     (1.27 µs … 1.32 µs)    1.3 µs   1.32 µs   1.32 µs
```

8 headers
```
benchmark             time (avg)        iter/s             (min … max)       p75       p99      p995
---------------------------------------------------------------------- -----------------------------
Headers iterator       9.78 µs/iter     102,259.9    (9.13 µs … 97.81 µs)   9.57 µs  12.61 µs   50.1 µ
```